+++ packages/mobx-state-tree/__tests__/core/api.test.ts
@@ -20,7 +20,8 @@ const METHODS_AND_INTERNAL_TYPES = stringToArray(`
     onAction,
     recordActions,
     createActionTrackingMiddleware,
-    setLivelynessChecking,
+    setLivelinessChecking,
+    getLivelinessChecking,
     getType,
     getChildType,
     onPatch,
@@ -76,6 +77,10 @@ const METHODS_AND_INTERNAL_TYPES = stringToArray(`
     types
 `)
 
+const DEPRECATED_METHODS_AND_INTERNAL_TYPES = stringToArray(`
+    setLivelynessChecking,
+`)
+
 const METHODS = METHODS_AND_INTERNAL_TYPES.filter(s => s[0].toLowerCase() === s[0])
 
 const INTERNAL_TYPES = METHODS_AND_INTERNAL_TYPES.filter(s => s[0].toUpperCase() === s[0])
@@ -113,6 +118,7 @@ test("correct api exposed", () => {
         Object.keys(mst)
             .sort()
             .filter(key => (mst as any)[key] !== undefined) // filter out interfaces
+            .filter(s => !DEPRECATED_METHODS_AND_INTERNAL_TYPES.includes(s))
     ).toEqual([...METHODS, ...INTERNAL_TYPES].sort())
 })
 test("correct types exposed", () => {
@@ -123,6 +129,7 @@ test("all methods mentioned in readme.md", () => {
     const missing = TYPES.map(type => "types." + type)
         .concat(METHODS)
         .filter(identifier => readme.indexOf("`" + identifier) === -1)
+        .filter(s => !DEPRECATED_METHODS_AND_INTERNAL_TYPES.includes(s))
     expect(missing).toEqual([])
 })
 test("all methods mentioned in api.md", () => {