+++ packages/mobx-state-tree/src/types/complex-types/model.ts
@@ -50,22 +50,12 @@ import {
     TypeFlags,
     ExtractC,
     ExtractS,
-    ExtractT
+    ExtractT,
+    Hook
 } from "../../internal"
 
 const PRE_PROCESS_SNAPSHOT = "preProcessSnapshot"
 const POST_PROCESS_SNAPSHOT = "postProcessSnapshot"
-
-/**
- * @internal
- * @private
- */
-export enum HookNames {
-    afterCreate = "afterCreate",
-    afterAttach = "afterAttach",
-    beforeDetach = "beforeDetach",
-    beforeDestroy = "beforeDestroy"
-}
 
 export interface ModelProperties {
     [key: string]: IAnyType
@@ -240,7 +230,7 @@ function toPropertiesObject(declaredProps: ModelPropertiesDeclaration): ModelPro
     return Object.keys(declaredProps).reduce(
         (props, key) => {
             // warn if user intended a HOOK
-            if (key in HookNames)
+            if (key in Hook)
                 return fail(
                     `Hook '${key}' was defined as property. Hooks should be defined as part of the actions`
                 )
@@ -379,7 +369,7 @@ export class ModelType<P extends ModelProperties, O> extends ComplexType<any, an
 
             // apply hook composition
             let baseAction = (self as any)[name]
-            if (name in HookNames && baseAction) {
+            if (name in Hook && baseAction) {
                 let specializedAction = action2
                 action2 = function() {
                     baseAction.apply(null, arguments)